Refactor operations functions to static methods
