+++ libraries/mult_submits.lib.php
@@ -8,10 +8,12 @@
  * @package PhpMyAdmin
  */
 
+use PhpMyAdmin\Operations;
 use PhpMyAdmin\Sql;
 use PhpMyAdmin\Table;
 use PhpMyAdmin\Transformations;
 use PhpMyAdmin\Url;
+use PhpMyAdmin\Util;
 
 /**
  * Gets url params
@@ -47,7 +49,7 @@ function PMA_getUrlParams(
     foreach ($selected as $sval) {
         if ($what == 'row_delete') {
             $_url_params['selected'][] = 'DELETE FROM '
-                . PhpMyAdmin\Util::backquote($table)
+                . Util::backquote($table)
                 . ' WHERE ' . $sval . ' LIMIT 1;';
         } else {
             $_url_params['selected'][] = $sval;
@@ -116,7 +118,7 @@ function PMA_buildOrExecuteQueryForMulti(
         case 'drop_db':
             PMA_relationsCleanupDatabase($selected[$i]);
             $a_query   = 'DROP DATABASE '
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $reload    = 1;
             $run_parts = true;
             $rebuild_database_list = true;
@@ -127,104 +129,104 @@ function PMA_buildOrExecuteQueryForMulti(
             $current = $selected[$i];
             if (!empty($views) && in_array($current, $views)) {
                 $sql_query_views .= (empty($sql_query_views) ? 'DROP VIEW ' : ', ')
-                          . PhpMyAdmin\Util::backquote($current);
+                          . Util::backquote($current);
             } else {
                 $sql_query .= (empty($sql_query) ? 'DROP TABLE ' : ', ')
-                           . PhpMyAdmin\Util::backquote($current);
+                           . Util::backquote($current);
             }
             $reload    = 1;
             break;
 
         case 'check_tbl':
             $sql_query .= (empty($sql_query) ? 'CHECK TABLE ' : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $execute_query_later = true;
             break;
 
         case 'optimize_tbl':
             $sql_query .= (empty($sql_query) ? 'OPTIMIZE TABLE ' : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $execute_query_later = true;
             break;
 
         case 'analyze_tbl':
             $sql_query .= (empty($sql_query) ? 'ANALYZE TABLE ' : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $execute_query_later = true;
             break;
 
         case 'checksum_tbl':
             $sql_query .= (empty($sql_query) ? 'CHECKSUM TABLE ' : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $execute_query_later = true;
             break;
 
         case 'repair_tbl':
             $sql_query .= (empty($sql_query) ? 'REPAIR TABLE ' : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i]);
+                       . Util::backquote($selected[$i]);
             $execute_query_later = true;
             break;
 
         case 'empty_tbl':
             $deletes = true;
             $a_query = 'TRUNCATE ';
-            $a_query .= PhpMyAdmin\Util::backquote($selected[$i]);
+            $a_query .= Util::backquote($selected[$i]);
             $run_parts = true;
             break;
 
         case 'drop_fld':
             PMA_relationsCleanupColumn($db, $table, $selected[$i]);
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                 : ',')
-                       . ' DROP ' . PhpMyAdmin\Util::backquote($selected[$i])
+                       . ' DROP ' . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ';' : '');
             break;
 
         case 'primary_fld':
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                     . (empty($primary)
                     ? ''
                     : ' DROP PRIMARY KEY,') . ' ADD PRIMARY KEY( '
                 : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i])
+                       . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ');' : '');
             break;
 
         case 'index_fld':
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                     . ' ADD INDEX( '
                 : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i])
+                       . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ');' : '');
             break;
 
         case 'unique_fld':
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                     . ' ADD UNIQUE( '
                 : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i])
+                       . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ');' : '');
             break;
 
         case 'spatial_fld':
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                     . ' ADD SPATIAL( '
                 : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i])
+                       . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ');' : '');
             break;
 
         case 'fulltext_fld':
             $sql_query .= (empty($sql_query)
-                ? 'ALTER TABLE ' . PhpMyAdmin\Util::backquote($table)
+                ? 'ALTER TABLE ' . Util::backquote($table)
                     . ' ADD FULLTEXT( '
                 : ', ')
-                       . PhpMyAdmin\Util::backquote($selected[$i])
+                       . Util::backquote($selected[$i])
                        . (($i == $selected_cnt-1) ? ');' : '');
             break;
 
@@ -232,9 +234,9 @@ function PMA_buildOrExecuteQueryForMulti(
             $newtablename = $_POST['add_prefix'] . $selected[$i];
             // ADD PREFIX TO TABLE NAME
             $a_query = 'ALTER TABLE '
-                . PhpMyAdmin\Util::backquote($selected[$i])
+                . Util::backquote($selected[$i])
                 . ' RENAME '
-                . PhpMyAdmin\Util::backquote($newtablename);
+                . Util::backquote($newtablename);
             $run_parts = true;
             break;
 
@@ -256,9 +258,9 @@ function PMA_buildOrExecuteQueryForMulti(
             }
             // CHANGE PREFIX PATTERN
             $a_query = 'ALTER TABLE '
-                . PhpMyAdmin\Util::backquote($selected[$i])
+                . Util::backquote($selected[$i])
                 . ' RENAME '
-                . PhpMyAdmin\Util::backquote($newtablename);
+                . Util::backquote($newtablename);
             $run_parts = true;
             break;
 
@@ -282,8 +284,7 @@ function PMA_buildOrExecuteQueryForMulti(
             $copy_tbl = true;
             Table::moveCopy($db, $selected[$i], $_POST['target_db'], $selected[$i], $_POST['what'], false, 'one_table');
             if (isset($_POST['adjust_privileges']) && !empty($_POST['adjust_privileges'])) {
-                include_once 'operations.lib.php';
-                PMA_AdjustPrivileges_copyTable($db, $selected[$i], $_POST['target_db'], $selected[$i]);
+                Operations::adjustPrivilegesCopyTable($db, $selected[$i], $_POST['target_db'], $selected[$i]);
             }
             break;
         } // end switch
@@ -454,7 +455,7 @@ function PMA_getHtmlForOtherActions($what, $action, $_url_params, $full_query)
     // Display option to disable foreign key checks while dropping tables
     if ($what === 'drop_tbl' || $what === 'empty_tbl' || $what === 'row_delete') {
         $html .= '<div id="foreignkeychk">';
-        $html .= PhpMyAdmin\Util::getFKCheckbox();
+        $html .= Util::getFKCheckbox();
         $html .= '</div>';
     }
     $html .= '<input id="buttonYes" type="submit" name="mult_btn" value="'
@@ -493,7 +494,7 @@ function PMA_getQueryFromSelected($what, $table, $selected, $views)
         switch ($what) {
         case 'row_delete':
             $full_query .= 'DELETE FROM '
-                . PhpMyAdmin\Util::backquote(htmlspecialchars($table))
+                . Util::backquote(htmlspecialchars($table))
                 // Do not append a "LIMIT 1" clause here
                 // (it's not binlog friendly).
                 // We don't need the clause because the calling panel permits
@@ -503,7 +504,7 @@ function PMA_getQueryFromSelected($what, $table, $selected, $views)
             break;
         case 'drop_db':
             $full_query .= 'DROP DATABASE '
-                . PhpMyAdmin\Util::backquote(htmlspecialchars($sval))
+                . Util::backquote(htmlspecialchars($sval))
                 . ';<br />';
             $reload = true;
             break;
@@ -512,31 +513,31 @@ function PMA_getQueryFromSelected($what, $table, $selected, $views)
             $current = $sval;
             if (!empty($views) && in_array($current, $views)) {
                 $full_query_views .= (empty($full_query_views) ? 'DROP VIEW ' : ', ')
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($current));
+                    . Util::backquote(htmlspecialchars($current));
             } else {
                 $full_query .= (empty($full_query) ? 'DROP TABLE ' : ', ')
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($current));
+                    . Util::backquote(htmlspecialchars($current));
             }
             break;
 
         case 'empty_tbl':
             $full_query .= 'TRUNCATE ';
-            $full_query .= PhpMyAdmin\Util::backquote(htmlspecialchars($sval))
+            $full_query .= Util::backquote(htmlspecialchars($sval))
                         . ';<br />';
             break;
 
         case 'primary_fld':
             if ($full_query == '') {
                 $full_query .= 'ALTER TABLE '
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($table))
+                    . Util::backquote(htmlspecialchars($table))
                     . '<br />&nbsp;&nbsp;DROP PRIMARY KEY,'
                     . '<br />&nbsp;&nbsp; ADD PRIMARY KEY('
                     . '<br />&nbsp;&nbsp;&nbsp;&nbsp; '
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($sval))
+                    . Util::backquote(htmlspecialchars($sval))
                     . ',';
             } else {
                 $full_query .= '<br />&nbsp;&nbsp;&nbsp;&nbsp; '
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($sval))
+                    . Util::backquote(htmlspecialchars($sval))
                     . ',';
             }
             if ($i == $selected_cnt-1) {
@@ -547,10 +548,10 @@ function PMA_getQueryFromSelected($what, $table, $selected, $views)
         case 'drop_fld':
             if ($full_query == '') {
                 $full_query .= 'ALTER TABLE '
-                    . PhpMyAdmin\Util::backquote(htmlspecialchars($table));
+                    . Util::backquote(htmlspecialchars($table));
             }
             $full_query .= '<br />&nbsp;&nbsp;DROP '
-                . PhpMyAdmin\Util::backquote(htmlspecialchars($sval))
+                . Util::backquote(htmlspecialchars($sval))
                 . ',';
             if ($i == $selected_cnt - 1) {
                 $full_query = preg_replace('@,$@', ';<br />', $full_query);
